sounds_good_enough/backend/pyproject.toml
[project]
name = "sounds-good-enough-backend"
version = "0.1.0"
description = "Backend for Sounds Good Enough"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "demucs>=4.0.1",
  "torchcodec",
  "soundfile",
  "gradium>=0.5.7",
  "yt-dlp>=2025.1.26",
]

[dependency-groups]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.27.2",
  "mypy>=1.11.2",
]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_configs = true
exclude = ["^tests/"]

[[tool.mypy.overrides]]
module = ["gradium", "gradium.*"]
ignore_missing_imports = true

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
